Use CrawlSpider with rules to scrape links
CrawlSpider provides an easy way to follow links that the spider crawls
to. The CrawlerProcess object defines settings to be used by the
crawler. Some settings include the data format and the output file in
which the data will be written to.

The CLOSESPIDER_ITEMCOUNT setting defines the maximum number of pages in
which the crawler follows links.

It also helps in running Scrapy from a regular Python script, instead of
using the 'scrapy' command:

python main.py [--max]

Default value of max is 10. It sets the CLOSESPIDER_ITEMCOUNT in the
CrawlerProcess, which defines the maximum number of pages in which the
crawler follows links.

Also deleted most lines in the settings.py file since we define our
desired settings in the CrawlerProcess object.

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>