Create helper file
Helper file contains objects we might often import, like the stemmer,
stopwords, etc. Also has a helper function in which we can pass in a
string of text and obtain a clean list of terms from the text. Makes for
less cluttered code.

Made it so that content from pages don't get stemmed with the
PorterStemmer. This is because that highly likely affect the AFINN
sentiment values of terms.

Also updated README.md.

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>