Implement Query class, makes AND queries
When the index is built, the user can decide if they want to conduct
queries against it. For each query, the user gets returned a list of
pages from the list of scraped pages in which each term in the query
appears.

Kept stopwords in page content since it was providing inconsistent
results.

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>